pyproject.toml
[project]
name = "pingme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioredis==2.0.1",
    "aiosmtplib==4.0.0",
    "alembic==1.14.1",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "async-timeout==5.0.1",
    "asyncpg==0.30.0",
    "authlib==1.4.1",
    "bcrypt==4.2.1",
    "black==25.1.0",
    "boto3>=1.41.5",
    "boto3-stubs==1.42.4",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "contextvars==2.4",
    "cryptography==44.0.0",
    "dnspython==2.7.0",
    "ecdsa==0.19.0",
    "email-validator==2.2.0",
    "fastapi==0.115.8",
    "firebase-admin>=7.1.0",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "hiredis==3.1.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "immutables==0.21",
    "mako==1.3.8",
    "markupsafe==3.0.2",
    "mypy-extensions==1.0.0",
    "packaging==24.2",
    "passlib==1.7.4",
    "pathspec==0.12.1",
    "platformdirs==4.3.6",
    "psutil==7.0.0",
    "psycopg2-binary>=2.9.11",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart>=0.0.20",
    "redis==5.2.1",
    "rsa==4.9",
    "ruff>=0.14.8",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.37",
    "starlette==0.45.3",
    "structlog>=25.4.0",
    "typing-extensions==4.12.2",
    "uvicorn[standard]==0.34.3",
    "websockets==15.0.1",
]

[tool.ruff]
target-version = "py313"
line-length = 120
exclude = [
  ".venv",
]
format.quote-style = "single"
lint.select = [
  "F", # Pyflakes
  # "E/W", # pycodestyle
  "C90", # mccabe
  # "I", # isort
  # "N", # pep8-naming
  # "D", # pydocstyle
  "UP", # pyupgrade
  "YTT", # flake8-2020
  # "ANN", # flake8-annotations
  # "ASYNC", # flake8-async
  # "TRIO", # flake8-trio
  # "S", # flake8-bandit
  # "BLE", # flake8-blind-except
  # "FBT", # flake8-boolean-trap
  "B", # flake8-bugbear
  # "A", # flake8-builtins
  # "COM", # flake8-commas
  # "CPY", # flake8-copyright
  "C4", # flake8-comprehensions
  # "DTZ", # flake8-datetimez
  "T10", # flake8-debugger
  # "DJ", # flake8-django
  # "EM", # flake8-errmsg
  # "EXE", # flake8-executable
  # "FA", # flake8-future-annotations
  "ISC", # flake8-implicit-str-concat
  # "ICN", # flake8-import-conventions
  "G", # flake8-logging-format
  "INP", # flake8-no-pep420
  # "PIE", # flake8-pie
  "T20", # flake8-print
  # "PYI", # flake8-pyi
  # "PT", # flake8-pytest-style
  # "Q", # flake8-quotes
  "RSE", # flake8-raise
  "RET", # flake8-return
  "SLF", # flake8-self
  # "SLOT", # flake8-slots
  "SIM", # flake8-simplify
  "TID", # flake8-tidy-imports
  # "TCH", # flake8-type-checking
  # "INT", # flake8-gettext
  # "ARG", # flake8-unused-arguments
  "PTH", # flake8-use-pathlib
  # "TD", # flake8-todos
  # "FIX", # flake8-fixme
  "ERA", # eradicate
  # "PD", # pandas-vet
  # "PGH", # pygrep-hooks
  "PL", # Pylint
  # "TRY", # tryceratops
  # "FLY", # flynt
  # "NPY", # NumPy-specific rules
  # "AIR", # Airflow
  # "PERF", # Perflint
  # "FURB", # refurb
  # "LOG", # flake8-logging
  "RUF" # Ruff-specific rules
]
lint.flake8-tidy-imports.ban-relative-imports = "all"
lint.flake8-bugbear.extend-immutable-calls = [
  "fastapi.Depends",
  "fastapi.params.Depends",
  "fastapi.Query",
  "fastapi.params.Query"
]
lint.mccabe.max-complexity = 20
